requirements.txt
scipy
matplotlib
ipython
jupyter
pandas
tqdm
scikit-learn
tensorflow-macos
da-dapper
